+++ src/rpc/rawtransaction.cpp
@@ -3,7 +3,6 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "base58.h"
 #include "chain.h"
 #include "coins.h"
 #include "consensus/validation.h"
@@ -11,6 +10,7 @@
 #include "init.h"
 #include "keystore.h"
 #include "validation.h"
+#include <key_io.h>
 #include "merkleblock.h"
 #include "net.h"
 #include "policy/policy.h"
@@ -740,16 +740,14 @@ UniValue signrawtransaction(const JSONRPCRequest& request)
         UniValue keys = request.params[2].get_array();
         for (unsigned int idx = 0; idx < keys.size(); idx++) {
             UniValue k = keys[idx];
-            CPAIcoinSecret vchSecret;
-            bool fGood = vchSecret.SetString(k.get_str());
-            if (!fGood)
-                throw JSONRPCError(RPCErrorCode::INVALID_ADDRESS_OR_KEY, "Invalid private key");
-            CKey key = vchSecret.GetKey();
-            if (!key.IsValid())
-                throw JSONRPCError(RPCErrorCode::INVALID_ADDRESS_OR_KEY, "Private key outside allowed range");
+            CKey key = DecodeSecret(k.get_str());
+            if (!key.IsValid()) {
+              throw JSONRPCError(RPCErrorCode::INVALID_ADDRESS_OR_KEY, "Invalid private key");
+            }
             tempKeystore.AddKey(key);
         }
     }
+
 #ifdef ENABLE_WALLET
     else if (pwallet) {
         EnsureWalletIsUnlocked(pwallet);