+++ src/wallet/rpcwallet.cpp
@@ -4,7 +4,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "amount.h"
-#include "base58.h"
+#include <key_io.h>
 #include "chain.h"
 #include "consensus/validation.h"
 #include "core_io.h"